+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [QA] Reorganize e2e tests around scopes
 - [REFACTOR] Introduce common LiveBlock component
 - [REFACTOR] Introduce common StaticBlock component
+- [FIX] Transform being applied to soon due to placeholder space character in new cells
 
 ## 0.10.1